created db functions for searching car listings and retrieving car by ID. Also added route for search listing
